requirements.txt
asgiref==3.5.0
boto3==1.21.42
botocore==1.24.42
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2
django-allauth==0.50.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
djangorestframework==3.13.0
djangorestframework-simplejwt==5.1.0
django-storages==1.12.3
idna==3.3
jmespath==1.0.0
oauthlib==3.2.0
Pillow==9.1.0
pipdeptree==2.2.1
psycopg2==2.9.3
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
